Add files via upload
This is the view of the UI of our 'Play' webpage. You will see all other buttons to other portions of the website still exist at the top of the website. There is a box where you can either use arrows or your keyboard to enter a bet amount and then press start game.