Add files via upload
This is our 'Earn' section of our website. We need a place to be able to give a break from blackjack as well as earn chips back if you reach 0. You click on collect trash and an animation plays and then you receive 100 chips. You can repeat this as many times as  you'd like.